Updated UML
Updated UML, now includes a password String in Bracket. In addition, ScoreboardPane now extends StackPane, and the login field in MarchMadnessGUI should extend FlowPane. Grant pointed out that in order to properly center a Pane in a BorderPane, it needs to have a Pos attribute, which the Pane class itself does not have. StackPane and FlowPane both have this attribute. It's reasonable to extend or use another Pane class, as long as it has this attribute and responds to it accordingly.